add edit page and form component
