+++ src/App.jsx
@@ -6,11 +6,12 @@ import {
 } from 'react-router-dom';
 import Navbar from './components/navbar/Navbar';
 import AboutScreen from './routes/About';
-import AdminScreen from './routes/Admin';
+import AdminScreen from './routes/admin/Admin';
 import BookingScreen from './routes/Booking';
 import DetailScreen from './routes/Detail';
+import EditScreen from './routes/admin/Edit';
 import HomeScreen from './routes/Home';
-import NewScreen from './routes/New';
+import NewScreen from './routes/admin/New';
 import ReservationScreen from './routes/Reservation';
 
 function App() {
@@ -20,6 +21,7 @@ function App() {
       <Routes>
         <Route element={<AdminScreen />} path="/admin" />
         <Route element={<NewScreen />} path="/admin/new" />
+        <Route element={<EditScreen />} path="/admin/edit/:id" />
         <Route element={<AboutScreen />} path="/about" />
         <Route element={<BookingScreen />} path="/booking" />
         <Route element={<DetailScreen />} path="/detail/:id" />
